LPD-31424 Create individual tab in filter attributes
